requirements.txt
fastapi
uvicorn
sqlalchemy
dotenv
psycopg2-binary
bcrypt
uuid
smtp-mail
python-jose[cryptography]